Juros correção, incompleta.

+++ Header Files/ContaCorrente.h
@@ -8,7 +8,7 @@ class ContaCorrente:public Conta{
       float taxaDeManutencao;
     public:
       //RETIRE AS CHAVES QUANDO FOR DEFINIR O CONSTRUTOR
-      ContaCorrente(const Usuario & = {"0",0,0,{"0","0","0","0"}},const int &  = 0,const float & = 0);
+      ContaCorrente(const Usuario & = {"0",0,0,{"0","0","0","0"}},const float &  = 0,const float & = 0);
       void setTaxaDeManutencao(const float &); //desconto para a manutenção da conta corrente (geralmente 15 reais por mes)
       void CalcularTaxaDeManutencao(const float &);
 };
