+++ Resource Files/ContaCorrente.cpp
@@ -1,13 +1,13 @@
 #include "..\Header Files\ContaCorrente.h"
 
 
-ContaCorrente::ContaCorrente(const Usuario & user,const int & taxaDeManutencao,const float & saldo):Conta(user,saldo){
-  this-> taxaDeManutencao=taxaDeManutencao;
+ContaCorrente::ContaCorrente(const Usuario & user,const float & taxaDeManutencao,const float & saldo):Conta(user,saldo){
+  this->taxaDeManutencao=taxaDeManutencao;
 }
 
 
-void ContaCorrente::setTaxaDeManutencao (const int & taxaDeManutencao){
-    this-> taxaDeManutencao = taxaDeManutencao;
+void ContaCorrente::setTaxaDeManutencao (const float & taxaDeManutencao){
+    this->taxaDeManutencao = taxaDeManutencao;
 }
 void ContaCorrente::CalcularTaxaDeManutencao(const float & taxaDeManutencao){
   this->taxaDeManutencao=taxaDeManutencao;
